migration: add xml dump import for series
* switch from importing multipart monographs and journals
  from json to xml dump (remove intermediate step)
