+++ cds_ils/importer/providers/cds/models/multipart.py
@@ -23,7 +23,8 @@ class CDSMultipart(CdsIlsOverdo):
     """Translation Index for CDS Books."""
 
     __query__ = (
-        "003:SzGeCERN 596__:MULTIVOLUMES -980__:DELETED -980__:MIGRATED"
+        "596__:MULTIVOLUMES "
+        "-980__:DELETED -980__:MIGRATED"
     )
 
     __ignore_keys__ = CDS_IGNORE_FIELDS
