+++ cds_ils/migrator/ill/api.py
@@ -54,16 +54,15 @@ import json
 
 import click
 from elasticsearch_dsl import Q
-from invenio_app_ils.ill.api import Library
 from invenio_app_ils.ill.proxies import current_ils_ill
 from invenio_db import db
 
-from cds_ils.migrator.api import bulk_index_records, import_record, \
-    model_provider_by_rectype
+from cds_ils.migrator.api import import_record
 from cds_ils.migrator.errors import BorrowingRequestError, ItemMigrationError
 from cds_ils.migrator.items.api import get_item_by_barcode
-from cds_ils.migrator.utils import get_acq_ill_notes, get_cost, get_date, \
-    get_migration_document_pid, get_patron_pid
+from cds_ils.migrator.utils import bulk_index_records, get_acq_ill_notes, \
+    get_cost, get_date, get_migration_document_pid, get_patron_pid, \
+    model_provider_by_rectype
 
 
 def get_status(record):
@@ -117,7 +116,7 @@ def get_type(record):
     return ill_type
 
 
-def migrate_to_ils(record):
+def clean_record_json(record):
     """Create a record for ILS."""
     document_pid = None
     try:
@@ -183,7 +182,7 @@ def import_ill_borrowing_requests_from_json(dump_file):
         ils_records = []
         for record in input_data:
             ils_record = import_record(
-                migrate_to_ils(record),
+                clean_record_json(record),
                 model,
                 provider,
                 legacy_id_key="legacy_id",
