+++ cds_ils/migrator/relations/api.py
@@ -17,6 +17,7 @@ from invenio_app_ils.records_relations.api import RecordRelationsParentChild, \
 from invenio_app_ils.relations.api import MULTIPART_MONOGRAPH_RELATION, \
     SERIAL_RELATION, Relation
 from invenio_db import db
+from invenio_pidstore.errors import PIDDoesNotExistError
 
 from cds_ils.literature.api import get_record_by_legacy_recid
 from cds_ils.migrator.documents.api import \
@@ -68,36 +69,39 @@ def migrate_siblings_relation():
             related_sibling = None
             try:
                 related_sibling = get_record_by_legacy_recid(
-                    document_class, relation["related_recid"])
-            except DocumentMigrationError as e:
+                    document_class, relation["related_recid"]
+                )
+            except PIDDoesNotExistError as e:
                 pass
 
             # try to find sibling in series
             if related_sibling is None:
                 try:
                     related_sibling = get_record_by_legacy_recid(
-                        series_class,
-                        relation["related_recid"])
-                except MultipartMigrationError as e:
+                        series_class, relation["related_recid"]
+                    )
+                except PIDDoesNotExistError as e:
                     continue
 
             # validate relation type
             relation_type = Relation.get_relation_by_name(
-                relation["relation_type"])
+                relation["relation_type"]
+            )
 
             # create relation
-            if related_sibling:
-                current_document_record = \
-                    document_class.get_record_by_pid(document.pid)
+            if related_sibling and related_sibling["pid"] != document.pid:
+                current_document_record = document_class.get_record_by_pid(
+                    document.pid
+                )
                 try:
-                    create_sibling_child_relation(current_document_record,
-                                                  related_sibling,
-                                                  relation_type=relation_type)
+                    create_sibling_child_relation(
+                        current_document_record,
+                        related_sibling,
+                        relation_type=relation_type,
+                    )
                     db.session.commit()
                 except RecordRelationsError as e:
-                    click.secho(
-                        e.description, fg="red"
-                    )
+                    click.secho(e.description, fg="red")
                     continue
 
 
@@ -105,7 +109,7 @@ def link_documents_and_serials():
     """Link documents/multiparts and serials."""
     document_class = current_app_ils.document_record_cls
     document_search = current_app_ils.document_search_cls()
-    series_class = current_app_ils.series_record_class
+    series_class = current_app_ils.series_record_cls
     series_search = current_app_ils.series_search_cls()
 
     def link_records_and_serial(record_cls, search):
@@ -126,7 +130,7 @@ def link_documents_and_serials():
     click.echo("Creating serial relations...")
     link_records_and_serial(
         document_class,
-        document_search.filter("term", _migration__has_serial=True)
+        document_search.filter("term", _migration__has_serial=True),
     )
     link_records_and_serial(
         series_class,
