+++ cds_ils/migrator/series/__init__.py
@@ -6,3 +6,12 @@
 # the terms of the MIT License; see LICENSE file for more details.
 
 """CDS-ILS migrator module."""
+from cds_dojson.overdo import OverdoBase
+
+serial_marc21 = OverdoBase(entry_point_models="cds_ils.migrator.serial_model")
+journal_marc21 = OverdoBase(
+    entry_point_models="cds_ils.migrator.journal_model"
+)
+mutlipart_marc21 = OverdoBase(
+    entry_point_models="cds_ils.migrator.multipart_model"
+)
