+++ tests/migrator/test_model_matcher.py
@@ -38,7 +38,9 @@ def test_marc21_matcher_books():
     not_match = {"foo": "bar"}
 
     models_entrypoint = "cds_ils.importer.models"
-    series_models_entrypoint = "cds_ils.importer.series_models"
+    serial_models_entrypoint = "cds_ils.migrator.serial_model"
+    journal_models_entrypoint = "cds_ils.migrator.journal_model"
+    multipart_models_entrypoint = "cds_ils.migrator.multipart_model"
     assert book.model == matcher(book_blob1, models_entrypoint)
     assert book.model == matcher(book_blob2, models_entrypoint)
 
@@ -46,14 +48,14 @@ def test_marc21_matcher_books():
         # English book club should not be matched
         assert book.model == matcher(book_blob3, models_entrypoint)
     assert standard.model == matcher(standard_blob1, models_entrypoint)
-    assert serial.model == matcher(serial_blob1, series_models_entrypoint)
+    assert serial.model == matcher(serial_blob1, serial_models_entrypoint)
     assert multipart.model == matcher(
-        multipart_blob1, series_models_entrypoint
+        multipart_blob1, multipart_models_entrypoint
     )
     assert multipart.model == matcher(
-        multipart_blob2, series_models_entrypoint
+        multipart_blob2, multipart_models_entrypoint
     )
-    assert journal.model == matcher(journal_blob, series_models_entrypoint)
+    assert journal.model == matcher(journal_blob, journal_models_entrypoint)
     # make sure that it won't match with any CDS record
     assert default.model == matcher(cds_blob2, models_entrypoint)
     assert default.model == matcher(not_match, models_entrypoint)
